[android] Separate test-related code from RunLoop
Move parts related to testing into a separate file.
